requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.8.0.post1
fastapi==0.112.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-text-splitters==0.2.2
langsmith==0.1.108
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pypdfium2==4.30.0
pyphen==0.16.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
setuptools==74.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
sympy==1.13.2
tenacity==8.5.0
textstat==0.7.4
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.7
